Remove doc comment condensation from logical expressions
