+++ src/parser/src/transforms/logical-expressions/index.ts
@@ -1,0 +1,4 @@
+export {
+    applyLogicalExpressionCondensation,
+    type CondenseLogicalExpressionsOptions
+} from "./condensation.js";
